+++ CG/CG task3/Math/src/main/ru/vsu/cs/math/matrix/Search.java
@@ -5,10 +5,10 @@ public class Search {
      * Создание нового массива без заданной строки и столбца
      * Исходный массив не изменяется
      */
-    protected static double[][] matrixReduction(double[][] arr, int row, int col) {
+    protected static float[][] matrixReduction(float[][] arr, int row, int col) {
         int countRow = 0, countCol = 0;
         int len = arr.length;
-        double[][] det = new double[len - 1][len - 1];
+        float[][] det = new float[len - 1][len - 1];
 
         for (int i = 0; i < len; i++) {
             if (i == row) {
@@ -30,17 +30,17 @@ public class Search {
     /*
      * Детерминант массива любого размера
      */
-    protected static double detMatrix(double[][] arr) {
-        double sum = 0;
+    protected static float detMatrix(float[][] arr) {
+        float sum = 0;
         if (arr.length == 1) {
             return arr[0][0];
         } else if (arr.length == 2) {
             return arr[0][0] * arr[1][1] - arr[1][0] * arr[0][1];
         } else {
             for (int i = 0; i < arr.length; i++) {
-                double[][] arrMatrix = matrixReduction(arr, i, 0);
-                double arrDeterminant = detMatrix(arrMatrix);
-                double minor = (arr[i][0] * Math.pow(-1, i) * arrDeterminant);
+                float[][] arrMatrix = matrixReduction(arr, i, 0);
+                float arrDeterminant = detMatrix(arrMatrix);
+                float minor = (arr[i][0] * (float) Math.pow(-1, i) * arrDeterminant);
                 sum += minor;
             }
         }
