+++ CG/CG task3/Math/src/main/ru/vsu/cs/math/vector/IVector.java
@@ -5,19 +5,19 @@ public interface IVector<T extends IVector<T>> {
 
     T sub(T vector);
 
-    T mul(double scalar);
+    T mul(float scalar);
 
-    T div(double scalar);
+    T div(float scalar);
 
-    double length();
+    float length();
 
     T normalization();
 
-    double dotProduct(T vector);
+    float dotProduct(T vector);
 
     boolean equals(T vector);
 
     String toString();
 
-    double[][] getVector();
+    float[][] getVector();
 }
