+++ CG/CG task3/Math/src/main/ru/vsu/cs/math/vector/Vector3f.java
@@ -1,17 +1,17 @@
 package main.ru.vsu.cs.math.vector;
 
 public class Vector3f implements IVector<Vector3f>{
-    public double x;
-    public double y;
-    public double z;
+    public float x;
+    public float y;
+    public float z;
 
-    public Vector3f(double x, double y, double z) {
+    public Vector3f(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public Vector3f(double[] arr) {
+    public Vector3f(float[] arr) {
         if (arr.length != 3) {
             throw new ArithmeticException("Wrong array length to create vector");
         }
@@ -24,21 +24,21 @@ public class Vector3f implements IVector<Vector3f>{
      * Получение вектора в виде массива столбца
      */
     @Override
-    public double[][] getVector() {
-        return new double[][]{{x}, {y}, {z}};
+    public float[][] getVector() {
+        return new float[][]{{x}, {y}, {z}};
     }
 
     /*
      * Получение значения "X"
      */
-    public double getX() {
+    public float getX() {
         return x;
     }
 
     /*
      * Получение значения "Y"
      */
-    public double getY() {
+    public float getY() {
         return y;
     }
 
@@ -46,7 +46,7 @@ public class Vector3f implements IVector<Vector3f>{
      * Получение значения "Z"
      */
 
-    public double getZ() {
+    public float getZ() {
         return z;
     }
 
@@ -65,7 +65,7 @@ public class Vector3f implements IVector<Vector3f>{
      * Исходный вектор не изменяется
      * Возвращается новый вектор
      */
-    public Vector3f add(double x, double y, double z) {
+    public Vector3f add(float x, float y, float z) {
         return new Vector3f(this.x + x, this.y + y, this.z + z);
     }
 
@@ -84,7 +84,7 @@ public class Vector3f implements IVector<Vector3f>{
      * Исходный вектор не изменяется
      * Возвращается новый вектор
      */
-    public Vector3f sub(double x, double y, double z) {
+    public Vector3f sub(float x, float y, float z) {
         return new Vector3f(this.x - x, this.y - y, this.z - z);
     }
 
@@ -94,10 +94,10 @@ public class Vector3f implements IVector<Vector3f>{
      * Возвращается новый вектор
      */
     @Override
-    public Vector3f mul(double scalar) {
-        double x = this.x * scalar;
-        double y = this.y * scalar;
-        double z = this.z * scalar;
+    public Vector3f mul(float scalar) {
+        float x = this.x * scalar;
+        float y = this.y * scalar;
+        float z = this.z * scalar;
         return new Vector3f(x, y, z);
     }
 
@@ -107,13 +107,13 @@ public class Vector3f implements IVector<Vector3f>{
      * Возвращается новый вектор
      */
     @Override
-    public Vector3f div(double scalar) {
+    public Vector3f div(float scalar) {
         if (Math.abs(scalar) < 1e-14) {
             throw new ArithmeticException("You cant divide on 0");
         }
-        double x = this.x / scalar;
-        double y = this.y / scalar;
-        double z = this.z / scalar;
+        float x = this.x / scalar;
+        float y = this.y / scalar;
+        float z = this.z / scalar;
         return new Vector3f(x, y, z);
     }
 
@@ -121,8 +121,8 @@ public class Vector3f implements IVector<Vector3f>{
      * Получение длины вектора
      */
     @Override
-    public double length() {
-        return Math.sqrt(x * x + y * y + z * z);
+    public float length() {
+        return (float) Math.sqrt(x * x + y * y + z * z);
     }
 
     /*
@@ -130,7 +130,7 @@ public class Vector3f implements IVector<Vector3f>{
      */
     @Override
     public Vector3f normalization() {
-        double len = this.length();
+        float len = this.length();
         if (Math.abs(len) < 1e-14) {
             throw new ArithmeticException("This vector cannot be normalized because its length is zero");
         }
@@ -141,7 +141,7 @@ public class Vector3f implements IVector<Vector3f>{
      * Скалярное произведение векторов
      */
     @Override
-    public double dotProduct(Vector3f v) {
+    public float dotProduct(Vector3f v) {
         return this.x * v.x + this.y * v.y + this.z * v.z;
     }
 
@@ -149,9 +149,9 @@ public class Vector3f implements IVector<Vector3f>{
      * Векторное произведение
      */
     public Vector3f vectorProduct(Vector3f v) {
-        double x = this.y * v.z - this.z * v.y;
-        double y = this.z * v.x - this.x * v.z;
-        double z = this.x * v.y - this.y * v.x;
+        float x = this.y * v.z - this.z * v.y;
+        float y = this.z * v.x - this.x * v.z;
+        float z = this.x * v.y - this.y * v.x;
         return new Vector3f(x, y, z);
     }
